updates to api translation plugins
