+++ pkg/plugins/common/provider/provider.go
@@ -17,9 +17,9 @@ limitations under the License.
 package provider
 
 const (
-	OpenAI            = "openai"
-	Anthropic         = "anthropic"
-	AzureOpenAI       = "azure-openai"
-	Vertex            = "vertex"
-	AWSBedrockOpenAI  = "awsbedrock-openai"
+	OpenAI        = "openai"
+	Anthropic     = "anthropic"
+	AzureOpenAI   = "azure-openai"
+	Vertex        = "vertex"
+	BedrockOpenAI = "bedrock-openai"
 )
